requirements.txt
# LangChain
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-postgres
langchain-openai
langchain-google-genai; extra == "gemini"

# PDF
pypdf

# Banco de dados
psycopg[binary]
psycopg2-binary
pgvector
SQLAlchemy

# Providers
openai
google-generativeai

# Utilitários
python-dotenv
pydantic
pydantic-settings